requirements.txt
numpy
Pillow
opencv-python
nuscenes-devkit
matplotlib
